+++ base/Root/Module/ItemDefinition/PropertyDefinition/sql/text.ts
@@ -68,7 +68,7 @@ export function textSQLIn(arg: ISQLInInfo) {
       // as well as iframe content
       // and they might not be cleaned in the client side during setting since values tend to be fairly hot during
       // retrival, so they are cleaned here as well
-      FORBID_ATTR: ["sizes", "srcset", "src", "data-src", "loading"],
+      FORBID_ATTR: ["sizes", "srcset", "src", "data-src", "loading", "href"],
     });
   }
