+++ client/internal/text/serializer/custom.ts
@@ -1,0 +1,69 @@
+import React from "react";
+import { RichElement, ISerializationRegistryType } from ".";
+import { serializeElementBase, deserializeElementBase, deserializeElement, IElementBase, reactifyElementBase } from "./base";
+
+export function registerCustom(registry: ISerializationRegistryType) {
+  const boundDeserializeElement = deserializeElement.bind(null, registry);
+
+  function serializeCustom(custom: ICustom) {
+    return serializeElementBase(
+      registry,
+      custom,
+      "div",
+      "custom-" + custom.customType,
+      null,
+      custom.children,
+    );
+  }
+  
+  function deserializeCustom(node: HTMLDivElement): ICustom {
+    const base = deserializeElementBase(node);
+    let customType: string = null;
+    node.classList.forEach((c) => {
+      if (c.startsWith("custom-")) {
+        customType = c.substr(7);
+      }
+    });
+    const custom: ICustom = {
+      ...base,
+      type: "custom",
+      customType,
+      children: Array.from(node.childNodes).map(boundDeserializeElement).filter((n) => n !== null) as RichElement[],
+    }
+    return custom;
+  }
+
+  function reactifyCustom(custom: ICustom, customProps?: React.DetailedHTMLProps<React.HTMLAttributes<HTMLElement>, HTMLElement>,) {
+    return reactifyElementBase(
+      registry,
+      custom,
+      "div",
+      "custom-" + custom.customType,
+      null,
+      customProps,
+      custom.children,
+    );
+  }
+
+  registry.REACTIFY.custom = reactifyCustom;
+  registry.SERIALIZE.custom = serializeCustom;
+  registry.DESERIALIZE.byClassNamePrefix.custom = deserializeCustom;
+}
+
+/**
+ * The custom type represents a custom- element
+ */
+export interface ICustom extends IElementBase {
+  /**
+   * The type as custom
+   */
+  type: "custom";
+  /**
+   * Specifies which custom type it is
+   */
+  customType: string;
+  /**
+   * The children
+   */
+  children: RichElement[];
+}
